+++ Framework.Application/ICommand.cs
@@ -1,0 +1,7 @@
+﻿namespace Framework.Application
+{
+    public interface ICommand
+    {
+
+    }
+}
