+++ Framework.Application/ICommandHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace Framework.Application
+{
+    public interface ICommandHandler<TCommand> where TCommand : ICommand
+    {
+        void Handle(TCommand command);
+    }
+}
