+++ Framework.Application/ILoggableCommand.cs
@@ -1,0 +1,6 @@
+﻿using Framework.Application;
+
+public interface ILoggableCommand: ICommand
+{
+
+}
