+++ Framework.Application/IRetryableCommand.cs
@@ -1,0 +1,7 @@
+﻿using Framework.Application;
+
+public interface IRetryableCommand: ICommand
+{
+    int RetryCount { get; set; }
+    int RetryDuration { get; set; }
+}
